Fix major performance issue with JsonStringToMessage().
This is done by changing usage pattern of StringOutputStream. Current usage pattern of loop { Next(); Backup(); } of StringOutputStream is really slow outside Google. This is because STLStringResizeUninitialized() is un-optimized (it has to zero out the re-sized bytes). And STLStringResizeUninitialized() is invoked everytime Next() is invoked.

We change the usage pattern of StringOutputStream by maintaining a handle to the returned buffer and doing buffer management ourselves instead of calling Backup().